requirements.txt
altgraph==0.17.4
aspose-pdf==24.7.0
beautifulsoup4==4.12.3
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
comtypes==1.4.9
cryptography==43.0.0
docx2pdf==0.1.8
fuzzywuzzy==0.18.0
Levenshtein==0.25.1
lxml==5.2.2
numpy==2.0.1
packaging==24.1
pdfminer.six==20231228
pdfplumber==0.11.3
pefile==2024.8.26
pillow==10.4.0
pycparser==2.22
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pypdfium2==4.30.0
pytesseract==0.3.10
python-docx==1.1.2
python-Levenshtein==0.25.1
pywin32==308
pywin32-ctypes==0.2.3
rapidfuzz==3.9.7
soupsieve==2.6
tqdm==4.66.5
typing_extensions==4.12.2
